[Fix]: 수면 점수 자료형 float으로 수정(AI전달 데이터 페이지 참고)

+++ src/main/java/com/example/peakly/domain/dailySleep/dto/request/DailySleepLogRequest.java
@@ -18,5 +18,5 @@ public record DailySleepLogRequest(
         @NotNull(message = "수면의 질을 체크해주세요")
         @Min(value = 1, message = "수면 점수는 1 이상이어야 합니다")
         @Max(value = 5, message = "수면 점수는 5 이하여야 합니다")
-        Integer sleepScore
+        Float sleepScore
         ) {}
