+++ src/main/java/com/example/peakly/domain/dailySleep/dto/request/SleepLogUpdateByTimeRequest.java
@@ -18,6 +18,6 @@ public record SleepLogUpdateByTimeRequest(
         @NotNull(message = "수면의 질을 체크해주세요")
         @Min(value = 1, message = "수면 점수는 1 이상이어야 합니다")
         @Max(value = 5, message = "수면 점수는 5 이하여야 합니다")
-        Integer sleepScore
+        Float sleepScore
 ) {
 }
